requirements.txt
fastapi~=0.112.2
uvicorn~=0.32.1
openai~=0.28.1
google-auth
google-auth-oauthlib
google-api-python-client~=2.169.0
python-dotenv

h11~=0.16.0
jupyter~=1.0.0
pip~=23.2.1
pathlib~=1.0.1
panel~=1.2.3
attrs~=22.1.0
zlib~=1.2.13
distro~=1.9.0
wheel~=0.38.4
openssl~=3.0.13
cryptography~=41.0.3
pillow~=10.4.0
docutils~=0.18.1
sphinx~=5.0.2
tornado~=6.3.2
jinja2~=3.1.2
rsa~=4.9
pyasn1~=0.4.8
future~=0.18.3
requests~=2.31.0
dask~=2023.6.0
rich~=13.7.1
keras~=3.4.1
colorama~=0.4.6
ipywidgets~=8.0.4
ipython~=8.15.0
notebook~=6.5.4
typing_extensions~=4.12.2
sniffio~=1.2.0
pytest~=7.4.0
click~=8.0.4
pygments~=2.15.1
httpcore~=1.0.9
idna~=3.4
httpx~=0.28.1
brotli~=1.0.9
zstandard~=0.19.0
aiohttp~=3.8.5
cachetools~=5.5.0
certifi~=2025.4.26
urllib3~=1.26.16
packaging~=23.1
cffi~=1.15.1
pydantic~=1.10.8
responses~=0.13.3
anyio~=3.5.0
numpy~=1.26.4
pandas~=2.0.3
starlette~=0.38.2
ujson~=5.4.0
pyparsing~=3.0.9
httplib2~=0.22.0
oauthlib~=3.2.2
deprecated~=1.2.14
chardet~=4.0.0
itsdangerous~=2.0.1
yaml~=0.2.5
pyyaml~=6.0
setuptools~=68.0.0
zipp~=3.11.0
build~=1.2.1
uritemplate~=4.1.1